+++ src/onegov.org/HISTORY.rst
@@ -1,6 +1,10 @@
 Changelog
 ---------
 
+- Gives users the ability to get a copy of their data by e-mail when submitting
+  a form or a reservation.
+  [href]
+
 - Adds the ability to group forms hierarchically.
   [href]